Create DataList component for History and savedBoards
