+++ src/app/[locale]/dashboard/@history/HistoryList.tsx
@@ -1,25 +1,9 @@
-'use client';
+'use client'; // Check avoid using use optimistic. and use as server component
 
-import { useState } from 'react';
-import Stack from '@mui/material/Stack';
-import List from '@mui/material/List';
-import Pagination from '@mui/material/Pagination';
 import { useOptimistic } from 'react';
-import DataItem from '@/components/DataItem/DataItem';
 import { removeHistoryData } from './actions';
 import { HistoryData } from './actions';
-
-const sxStyles = {
-  stack: {
-    display: 'flex',
-    flexDirection: 'column',
-    justifyContent: 'center',
-    alignItems: 'center',
-    width: '100%',
-  },
-  list: { width: '100%' },
-  pagination: { paddingBottom: 1 },
-};
+import DataList from '@/components/DataList/DataList';
 
 export default function HistoryList({
   initialHistories,
@@ -37,30 +21,5 @@ export default function HistoryList({
     await removeHistoryData(historyToDelete);
   };
 
-  const [page, setPage] = useState(1);
-  const handleChange = (event: React.ChangeEvent<unknown>, value: number) => {
-    setPage(value);
-  };
-  const itemsPerPage = 2;
-  const paginationCount = Math.ceil(histories.length / itemsPerPage);
-  const start = (page - 1) * itemsPerPage;
-  const truncatedItems = histories.slice(start, start + itemsPerPage);
-
-  return (
-    <Stack sx={sxStyles.stack}>
-      <List sx={sxStyles.list}>
-        {truncatedItems.map((data, index) => (
-          <DataItem data={data} key={index} onDelete={deleteHistoryData} />
-        ))}
-      </List>
-      <Pagination
-        sx={sxStyles.pagination}
-        count={paginationCount}
-        color="primary"
-        siblingCount={0}
-        page={page}
-        onChange={handleChange}
-      />
-    </Stack>
-  );
+  return <DataList list={histories} deleteItem={deleteHistoryData} />;
 }
